Add files via upload
The example for UnitTesting was too simple for what our group has done. Thought testing would be easier, but having onCreate,onStop,onStart etc. I did not understand fully how to test. It would be easy if it was a calculator with methods to push parameters through, but these did not.